+++ CHANGELOG.md
@@ -27,6 +27,9 @@
   * [Fixed] #819: ext-filter: Handle nodes without title.
   * [Fixed] #835: ext-dnd5: Accept drop externals after drag.
 
+# 'viewport' / Unreleased
+  * Render sub-parts of a tree
+  
 # 2.27.0 / 2017-12-16
   * **BREAKING CHANGES:**
     - `node.type` is now a first-class property of FancytreeNode.